Feature/tf2 baselink computation (#2)
Our base link isnt on the same rigid body as the lidar. this pr fixes that the tf from odom frame to our base link is published correctly in to tf

* fix issue with normalized imus

* Change how baselink transform is published

Instead of relying on static extrinsics to transform the odometry from
lidar to baselink, use tf2 lookup to query the transform between lidar
and baselink.

WARN: currently, only the transform is published for the baselink, the
odom is still in "lio" frame.

* port change from upstream

from
https://github.com/vectr-ucla/direct_lidar_inertial_odometry/pull/72

* "fix" runtime crash

apparently a race condition can happen if we run the ros publishing in a
detached thread which leads to a `free() invalid size` crash. This seems
to prevent the crash but is probably not the best solution.

* move transform publishing to pose publishing

so we get these things at a higher rate.
Add try catch to robustify node on startup.